pyproject.toml
[tool.poetry]
name = "pi4"
version = "0.1.0"
description = ""
authors = ["Afonso"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
bottle-dev = "^0.13.2"
scikit-learn = "^1.2.2"
pandas = "^2.0.1"
openpyxl = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
